requirements.txt
bidict==0.23.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
CounterFit==0.1.4.dev9
counterfit-connection==0.1.0.dev5
counterfit-shims-grove==0.1.4.dev5
counterfit-shims-seeed-python-dht==0.1.0.dev1
dnspython==2.6.1
eventlet==0.36.1
Flask==2.1.2
Flask-SocketIO==5.2.0
greenlet==3.1.0
h11==0.14.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
paho-mqtt==2.1.0
python-engineio==4.9.1
python-socketio==5.11.4
requests==2.32.3
setuptools==74.1.1
simple-websocket==1.0.0
six==1.16.0
urllib3==2.2.2
Werkzeug==2.0.0
wsproto==1.2.0
